backend/Rs485/requirements.txt
paho-mqtt
pymodbus<3.0.0
pyserial
python-dotenv

